+++ Kavita.Models/DTOs/UpdateSeriesDto.cs
@@ -21,7 +21,7 @@ public sealed record UpdateSeriesDto : IUpdateExternalMetadataIds
     public int? HardcoverId { get; set; }
     public long? MetronId { get; set; }
     public string? ComicVineId { get; set; }
-    public long? MangaBakaId { get; set; }
+    public int? MangaBakaId { get; set; }
     public int? CbrId { get; set; }
     #endregion
 }
